+++ packages/@webex/plugin-meetings/test/unit/spec/breakouts/breakout.ts
@@ -35,7 +35,7 @@ describe('plugin-meetings', () => {
     describe('#join', () => {
       it('makes the request as expected', async () => {
         const result = await breakout.join()
-  
+
         assert.calledOnceWithExactly(webex.request, {
           method: 'POST',
           uri: 'url/move',
@@ -44,7 +44,7 @@ describe('plugin-meetings', () => {
             sessionId: 'sessionId'
           }
         });
-  
+
         assert.equal(result, 'REQUEST_RETURN_VALUE')
       });
     });
@@ -88,7 +88,7 @@ describe('plugin-meetings', () => {
     describe('#askForHelp', () => {
       it('makes the request as expected', async () => {
         const result = await breakout.askForHelp()
-  
+
         assert.calledOnceWithExactly(webex.request, {
           method: 'POST',
           uri: 'url/help',
@@ -97,7 +97,35 @@ describe('plugin-meetings', () => {
             sessionId: 'sessionId'
           }
         });
-  
+
+        assert.equal(result, 'REQUEST_RETURN_VALUE')
+      });
+    });
+
+    describe('#broadcast', () => {
+      it('makes the request as expected', async () => {
+        breakout.breakoutRequest.broadcast = sinon.stub().returns(Promise.resolve('REQUEST_RETURN_VALUE'));
+        let result = await breakout.broadcast('hello')
+        assert.calledWithExactly(breakout.breakoutRequest.broadcast, {
+          url: 'url',
+          message: 'hello',
+          options: undefined,
+          groupId: 'groupId',
+          sessionId: 'sessionId'
+        });
+
+        assert.equal(result, 'REQUEST_RETURN_VALUE')
+
+        result = await breakout.broadcast('hello', {presenters: true, cohosts: true})
+
+        assert.calledWithExactly(breakout.breakoutRequest.broadcast, {
+          url: 'url',
+          message: 'hello',
+          options: {presenters: true, cohosts: true},
+          groupId: 'groupId',
+          sessionId: 'sessionId'
+        });
+
         assert.equal(result, 'REQUEST_RETURN_VALUE')
       });
     });